Added module and package deletion. - Reorganized menu bar and module pop-up menu, allowing them to diverge. - Added intSlotsCompare() to do bulk comparison in byte strings and   two-byte strings. - After parsing a monomorphic macro with no method definitions, it no   longer verifies the argument types, since they were checked during   parsing. - Documented LeftistHeap. - Message bundle trees are now easier to understand in the debugger. - generateTwoByteString now requires a generator producing a UShort. - added Strings utility truncateTo().

+++ avail/src/main/kotlin/avail/descriptor/tuples/NybbleTupleDescriptor.kt
@@ -210,7 +210,8 @@ class NybbleTupleDescriptor private constructor(
 					return when (val codePoint = strongNewElement.codePoint)
 					{
 						in 0 .. 0xFF -> generateByteString(1) { codePoint }
-						in 0 .. 0xFFFF -> generateTwoByteString(1) { codePoint }
+						in 0 .. 0xFFFF ->
+							generateTwoByteString(1) { codePoint.toUShort() }
 						else -> tuple(strongNewElement)
 					}
 				strongNewElement.isLong ->
